Implement TOO STEEP PATH marker on geometric path
